% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainrcode.R
\name{explainrcode}
\alias{explainrcode}
\alias{setting}
\title{Explain R code}
\usage{
setting()
}
\value{
A character string containing a natural language explanation of the R code.
}
\description{
Explain R code
}
\details{
This function takes a string of R code and uses GPT-3 to provide a natural language explanation
of what the code does.
}
