% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting.R
\name{setting}
\alias{setting}
\title{Set OpenAI API Key and Output Language}
\usage{
setting()
}
\description{
This function sets the OpenAI API key and output language to be used by other functions in the package.
The OpenAI API key is required for authentication when making requests to the OpenAI API. The output
language determines the language in which the results will be returned by the OpenAI API.
}
