% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttor.R
\name{texttor}
\alias{texttor}
\alias{setting}
\title{Convert Natural Language Text to R Code}
\usage{
setting()
}
\description{
Convert Natural Language Text to R Code
}
\details{
This function takes a string of natural language text as input and uses GPT-3 to generate
corresponding R code. The function returns a string of R code that corresponds to the
natural language input.
}
